Fix some (but not all) compiliation problems on windows

+++ src/windows/utility.cc
@@ -7,23 +7,23 @@ using namespace System;
 class WindowsUtility : public Utility
 {
 public:
-	virtual void Sleep(int milliseconds)
-	{
-		::Sleep(milliseconds);
-	}
-	
-	virtual int GetTicks()
-	{
-		return timeGetTime();
-	}
+    virtual void Sleep(int milliseconds)
+    {
+        ::Sleep(milliseconds);
+    }
+
+    virtual unsigned long GetTicks()
+    {
+        return timeGetTime();
+    }
 };
 
 Utility *Utility::GetInstance()
 {
-	static Utility *instance = NULL;
-	
-	if (instance == NULL)
-		instance = new WindowsUtility();
-	
-	return instance;
+    static Utility *instance = NULL;
+
+    if (instance == NULL)
+        instance = new WindowsUtility();
+
+    return instance;
 }
